+++ t2kdm/backends.py
@@ -1009,10 +1009,20 @@ class DIRACBackend(GridBackend):
         else:
             out = None
 
-        source = storage.get_SE(source_surl).name
-        destination = storage.get_SE(destination_surl).name
+        source = storage.get_SE(source_surl)
+        if source is None:
+            raise BackendException(
+                "Could not find storage element with host name string contained inside %s."
+                % (source_surl)
+            )
+        destination = storage.get_SE(destination_surl)
+        if destination is None:
+            raise BackendException(
+                "Could not find storage element with host name string contained inside %s."
+                % (destination_surl)
+            )
         try:
-            self._replicate_cmd(lurl, destination, source, _out=out, **kwargs)
+            self._replicate_cmd(lurl, destination.name, source.name, _out=out, **kwargs)
         except sh.ErrorReturnCode as e:
             if "No such file" in str(e.stderr):
                 raise DoesNotExistException("No such file or directory.")
@@ -1048,10 +1058,15 @@ class DIRACBackend(GridBackend):
             out = sys.stdout
         else:
             out = None
-        se = storage.get_SE(surl).name
+        se = storage.get_SE(surl)
+        if se is None:
+            raise BackendException(
+                "Could not find storage element with host name string contained inside %s."
+                % (surl)
+            )
 
         try:
-            self._add_cmd(lurl, localpath, se, _out=out, **kwargs)
+            self._add_cmd(lurl, localpath, se.name, _out=out, **kwargs)
         except sh.ErrorReturnCode as e:
             if "No such file" in str(e.stderr):
                 raise DoesNotExistException("No such file or directory.")
@@ -1063,7 +1078,12 @@ class DIRACBackend(GridBackend):
         return True
 
     def _remove(self, surl, lurl, last=False, verbose=False, **kwargs):
-        se = storage.get_SE(surl).name
+        se = storage.get_SE(surl)
+        if se is None:
+            raise BackendException(
+                "Could not find storage element with host name string contained inside %s."
+                % (surl)
+            )
 
         if last:
             # Delete lfn
@@ -1072,8 +1092,8 @@ class DIRACBackend(GridBackend):
             ret = self.dm.removeFile([lurl])
         else:
             if verbose:
-                print_("Removing replica of %s from %s." % (lurl, se))
-            ret = self.dm.removeReplica(se, [lurl])
+                print_("Removing replica of %s from %s." % (lurl, se.name))
+            ret = self.dm.removeReplica(se.name, [lurl])
 
         if not ret["OK"]:
             raise BackendException("Failed: %s" % (ret["Message"]))